Change to use RunTimeLibraries as CompileLibraries is empty for net6 after publishing
After updating to .net6 we realized that all of our attached links were no longer being added to responses.
Oddly enough when running locally (in debug) the links worked fine.
After some investigation we found that the routeMap was empty for the deployed system. This was because CompileLibraries was not populated.

For further context we publish with `dotnet publish XXX -c Release -r linux-x64 -v q`.

Switching to RunTimeLibraries works for both debugging and the published app. All tests still pass as expected.